requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
catboost==1.2.10
certifi==2026.2.25
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.135.1
fonttools==4.62.1
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
narwhals==2.18.0
numpy==2.4.2
openai==2.30.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pyarrow==24.0.0
pillow==12.1.1
plotly==6.6.0
pluggy==1.6.0
psycopg[binary]==3.3.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
xgboost==3.2.0
slowapi==0.1.9
PyJWT==2.12.1
cryptography==46.0.6
